% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtw.R
\name{dtw}
\alias{dtw}
\title{Dynamic Time Warping (DTW) with Traceback}
\usage{
dtw(x, y)
}
\arguments{
\item{x}{Numeric vector. The first sequence.}

\item{y}{Numeric vector. The second sequence.}
}
\value{
A list containing:
\describe{
  \item{DTW_matrix}{The cumulative DTW cost matrix.}
  \item{traceback_matrix}{A matrix storing the direction of the optimal path ("diag", "up", "left").}
  \item{path}{A list of coordinates representing the optimal alignment path.}
}
}
\description{
Computes the Dynamic Time Warping distance between two numeric sequences
and returns the DTW matrix, the traceback matrix, and the optimal alignment path.
}
\examples{
x <- c(1, 3, 4, 9)
y <- c(1, 3, 7, 8, 9)
dtw_result <- dtw(x, y)

}
